changed SlowTcpLink::id to a u32 from a u64

+++ src/tcp/tcp_junction.rs
@@ -263,7 +263,7 @@ impl SlowTcpJunction {
     ///
     /// # Arguments
     /// * `link_id` - The ID of the SlowTcpLink to remove
-    async fn remove_link(&self, link_id: u64) {
+    async fn remove_link(&self, link_id: u32) {
         let mut links = self.links.lock().await;
         // Find and remove the link with matching ID
         links.retain(|link| link.id() != link_id);
@@ -283,7 +283,7 @@ impl SlowTcpJunction {
     ///
     /// # Returns
     /// * `std::io::Result<usize>` - The number of bytes sent or an IO error
-    async fn broadcast(&self, data: &[u8], exclude_link_id: Option<u64>) -> std::io::Result<usize> {
+    async fn broadcast(&self, data: &[u8], exclude_link_id: Option<u32>) -> std::io::Result<usize> {
         // Get a reference to all active links
         let links = self.links.lock().await;
 
@@ -402,7 +402,7 @@ impl SlowTcpJunction {
     /// # Arguments
     /// * `data` - The slice of bytes received from the link
     /// * `link_id` - The ID of the link that received the data
-    async fn process(&self, data: &[u8], link_id: u64) {
+    async fn process(&self, data: &[u8], link_id: u32) {
         // Try to unpack the data into a SlowPackage
         let package = match SlowPackage::unpack(data) {
             Some(package) => package,
